Switch from threading to multiprocessing. Fixes #9.
The problem that was happening with a Thread would be that the main
process finished before the thread finished, which resulted in the
notifications not being sent.

A Process fixes this by allowing the notification to be sent before
exiting, since it uses an entirely different process.

This should have been using multiprocessing all along.