Add black formatter and isort
This commit mostly formats the files according to black's standards. It
also sorts imports according to isort.

A GitHub workflow was added to format on a push and a pre-commit config
was added to run formatting before a commit.

Lastly, a test in `test_account` was fixed because it was not working in
python 3.7