Remove `entrypoint.sh` in favor of passing in arguments through docker
Eventually I will add support for ENV variables within the script. I
wanted to have the way to run the script more universal and passing in
the arguments at the end of the command is exactly how it is done
without the use of a container.

This is how it can be run (will add to README):
docker run -d auto-southwest-check-in USERNAME PASSWORD