Add note about international flights to README
Fixes #21.
International flights will still prompt the user to fill out their
passport information before checking in if they haven't filled it out
yet. I'm still not sure whether or not this script works for
international flights. If another person files an issue similar to this,
I will change the note to say that the script only works with domestic
flights.