Start multiple accounts and flights in parallel
One flight/account is retrieved from the command line arguments, if
provided. The other flights/accounts will be read from the configuration
file.

Multiprocessing is used to achieve the parallelism. Due to this
parallelism, text printed to the screen could be out of order. However, I
haven't seen this happen yet.

I also cleaned up the mocks in some tests.