Add logging
Logging will make it much easier to debug. My goal is to not log any
sensitive information to make it easy for users to attach their logs.
This slightly limits the ability to debug as effectively by looking at
the logs, but the information that isn't logged isn't really needed to
debug in the first place.

There are multiple log files stored in the logs directory. They will
rollover whenever a new instance of the script is added or the size of
the file is above 2MB.