Build and push develop image only when necessary
Many files won't be in the container, so there is no point in running
the workflow when these are updated.

Also, I got rid of the tests workflow needing to be passed before
building the images because that will already be checked before making a
release. For the develop branch, it isn't necessary to have every test
pass before pushing the image.