Fix child processes not logging when process start method is 'spawn'
By default, the multiprocessing start method is 'fork' on Linux and
'spawn' on Windows and macOS. With the 'spawn' start method, the logging
configuration does not copy over to child processes, meaning only main
process log messages were actually logged.

Now, the start method is detected. If it is 'spawn', it will
reinitialize the logger configuration for that process. This enables the
logger to work the same on any start method used.

A drawback of logging this way (it was still a problem before this
commit) is that a race condition could be met if two loggers try to
write to the file/stdout at the same time. However, my testing shows
that one logger will wait for the other to be done. Therefore, no log
messages should be lost -- but if they are, it is very infrequent and
will most likely not cause problems debugging.