Reinitialize Chrome Options every time the webdriver is initialized
Fixes #98.

Selenium doesn't allow you to reuse the ChromeOptions object for
consecutive webdriver initializations. Therefore, get rid of the class
variable and simply call retrieve a new options object every time.

A test was modified to ensure this won't happen again.

A debug message was also modified to show information better (repr() only
works for exceptions that make a custom function for it. This new ways allows
the script to always log the exception info that is needed).