Don't use static methods in classes
All of these methods still go with the class even though they don't use
self. Additionally, they are not called outside of their class as a
static method. Therefore, it makes more logical sense to just include
them as part of the class instead of as a static method

I also deleted some unneeded mocks in the webdriver class due to the
get_options changes