Print ordered information about check-ins when Ctrl-C is pressed
An ordered list of all check-ins being cancelled will now be printed to
the screen instead of a Traceback when the script is interrupted is pressed.

Before each monitor stops, it sleeps for a certain amount of time (based
on its process name/number). This is done to ensure the information gets
printed ordered rather than in a jumbled mess.