Get rid of pop-ups on page loading when logging in
Fixes #226. For a reason unknown to me, Southwest's website now has a
pop-up saying "Something must have gone wrong. Please try again" when
running the Docker image (it is probably due to their website changes
that are happening now). Clicking "OK" on the pop-up removes it and
everything else works correctly, so that is the solution for now.