Remove timezone information when querying local time
datetime.now(timezone.utc) includes tzinfo whereas datetime.utcnow()
(which is now deprecated so it was changed to the former) does not
include tzinfo. This is a simple fix to just remove the tzinfo as there
is no timezone info for a flight's departure time (when comparing a time
with and without tzinfo, an exception would be thrown).

See https://github.com/jdholtz/auto-southwest-check-in/issues/256#issuecomment-2113020653
for a traceback of the issue before this fix.